+++ app/(Kambaz)/Courses/[cid]/Modules/page.tsx
@@ -2,20 +2,27 @@
 "use client";
 import { useParams } from "next/navigation";
 import React, { useState, useEffect } from 'react';
-// Replaced react-icons with inline SVG to resolve build error
 import { ListGroup, ListGroupItem, FormControl } from 'react-bootstrap';
 import LessonControlButtons from './LessonControlButtons'; 
 import ModuleControlButtons from './ModuleControlButtons';
 import ModulesControls from "./ModulesControls";
 import { addModule, editModule, updateModule, deleteModule, setModules } from "./reducer";
 import { useSelector, useDispatch } from "react-redux";
+import { RootState } from "../../../store"; 
 import * as client from "../../client"; 
 
 export default function Modules() {
   const { cid } = useParams();
-  const { modules } = useSelector((state: any) => state.modulesReducer);
+  
+  const { modules } = useSelector((state: RootState) => state.modulesReducer);
+  const { currentUser } = useSelector(
+    (state: RootState) => state.accountReducer as { currentUser: { role?: string, _id?: string } | null }
+  );
+  
   const dispatch = useDispatch();
   const [moduleName, setModuleName] = useState("");
+
+  const isFaculty = currentUser?.role === "FACULTY";
 
   const fetchModules = async () => {
     if (!cid) return; 
@@ -27,9 +34,8 @@ export default function Modules() {
     fetchModules();
   }, [cid]); 
 
-  // --- CRUD HANDLERS ---
   const onCreateModuleForCourse = async () => {
-    if (!cid) return;
+    if (!cid || !isFaculty) return; 
     const newModule = { name: moduleName, course: cid };
     const module = await client.createModuleForCourse((cid as string), newModule);
     dispatch(setModules([...modules, module]));
@@ -37,15 +43,13 @@ export default function Modules() {
   };
 
   const onRemoveModule = async (moduleId: string) => {
-    if (!cid) return;
-    // Pass cid to deleteModule
+    if (!cid || !isFaculty) return; 
     await client.deleteModule(cid as string, moduleId);
     dispatch(setModules(modules.filter((m: any) => m._id !== moduleId)));
   };
 
   const onUpdateModule = async (module: any) => {
-    if (!cid) return;
-    // Pass cid to updateModule
+    if (!cid || !isFaculty) return; 
     await client.updateModule(cid as string, module);
     const newModules = modules.map((m: any) => m._id === module._id ? module : m );
     dispatch(setModules(newModules));
@@ -53,23 +57,26 @@ export default function Modules() {
 
   return (
     <div id="wd-modules-screen">
-      <ModulesControls 
-        moduleName={moduleName} setModuleName={setModuleName}
-        addModule={onCreateModuleForCourse} 
-      />
+      {isFaculty && (
+        <ModulesControls 
+          moduleName={moduleName} setModuleName={setModuleName}
+          addModule={onCreateModuleForCourse} 
+        />
+      )}
       <br /><br /><br />
       
       <ListGroup id="wd-modules" className="rounded-0">
         {modules.map((module: any) =>(
           <ListGroupItem key={module._id} className="wd-module p-0 mb-5 fs-5 border-gray">
             <div className="wd-title p-3 ps-2 bg-secondary">
-              {/* Inline SVG for Grip Vertical Icon */}
-              <svg className="me-2 fs-3" xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
-                <path d="M7 2a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 8a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0z"/>
-              </svg>
               
-              {!module.editing && module.name}
-              { module.editing && (
+              {isFaculty && (
+                <svg className="me-2 fs-3" xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
+                  <path d="M7 2a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 8a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0z"/>
+                </svg>
+              )}
+              
+              {isFaculty && module.editing ? (
                 <FormControl 
                   className="w-50 d-inline-block"
                   defaultValue={module.name}
@@ -82,11 +89,15 @@ export default function Modules() {
                     }
                   }}
                 />
+              ) : (
+                module.name
               )}
+
               <ModuleControlButtons 
                 moduleId={module._id}
                 deleteModule={(moduleId) => onRemoveModule(moduleId)}
                 editModule={(moduleId) => dispatch(editModule(moduleId))}
+                isFaculty={isFaculty}
               /> 
             </div>
             
@@ -94,12 +105,13 @@ export default function Modules() {
               <ListGroup className="wd-lessons rounded-0">
                 {module.lessons.map((lesson: any) => (
                   <ListGroupItem key={lesson._id} className="wd-lesson p-3 ps-1">
-                    {/* Inline SVG for Grip Vertical Icon */}
-                    <svg className="me-2 fs-3" xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
-                      <path d="M7 2a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 8a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0z"/>
-                    </svg>
+                    {isFaculty && (
+                      <svg className="me-2 fs-3" xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
+                        <path d="M7 2a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zM7 8a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-3 3a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm3 0a1 1 0 1 1-2 0 1 1 0 0 1 2 0z"/>
+                      </svg>
+                    )}
                     {lesson.name} 
-                    <LessonControlButtons />
+                    <LessonControlButtons isFaculty={isFaculty} />
                   </ListGroupItem>
                 ))}
               </ListGroup>
